update : day24 (real16)

+++ test/24.f90
@@ -18,12 +18,12 @@ program main
    do i=1,n
       do j=i+1,n
          block
-            real(8)::a(2,2),b(2),work(2)
+            real(16)::a(2,2),b(2),work(2)
             integer::ipiv(2)
             !x1 + v1 * t1 = x2 + v2 * t2
             !v1 * t1 - v2 * t2 = x2 - x1
-            a(1,:)=[real(8)::v(1,i),-v(1,j)]! [t1]
-            a(2,:)=[real(8)::v(2,i),-v(2,j)]! [t2]
+            a(1,:)=[real(16)::v(1,i),-v(1,j)]! [t1]
+            a(2,:)=[real(16)::v(2,i),-v(2,j)]! [t2]
             b(1)=x(1,j)-x(1,i)
             b(2)=x(2,j)-x(2,i)
             if(inv(a, 2, work, ipiv))then
@@ -63,22 +63,22 @@ program main
       ! (0 ,-c,b )(0,y,z)
       ! (c , 0,-a)(x,0,z)
       ! (-b, a,0 )(x,y,0)
-      real(8)::a(6,6),b(6),work(6)
+      real(16)::a(6,6),b(6),work(6)
       integer::ipiv(6)
       associate(v1=>v(:,1:3),x1=>(x(:,1:3)))
          associate(v21=>v(:,2)-v(:,1),&
                v32=>v(:,3)-v(:,2),&
                x12=>x(:,1)-x(:,2),&
                x23=>x(:,2)-x(:,3))
-            a(1,:)=[real(8)::      0,-v21(3), v21(2),      0,-x12(3), x12(2)]
-            a(2,:)=[real(8):: v21(3),      0,-v21(1), x12(3),      0,-x12(1)]
-            a(3,:)=[real(8)::-v21(2), v21(1),      0,-x12(2), x12(1),      0]
-            a(4,:)=[real(8)::      0,-v32(3), v32(2),      0,-x23(3), x23(2)]
-            a(5,:)=[real(8):: v32(3),      0,-v32(1), x23(3),      0,-x23(1)]
-            a(6,:)=[real(8)::-v32(2), v32(1),      0,-x23(2), x23(1),      0]
+            a(1,:)=[real(16)::      0,-v21(3), v21(2),      0,-x12(3), x12(2)]
+            a(2,:)=[real(16):: v21(3),      0,-v21(1), x12(3),      0,-x12(1)]
+            a(3,:)=[real(16)::-v21(2), v21(1),      0,-x12(2), x12(1),      0]
+            a(4,:)=[real(16)::      0,-v32(3), v32(2),      0,-x23(3), x23(2)]
+            a(5,:)=[real(16):: v32(3),      0,-v32(1), x23(3),      0,-x23(1)]
+            a(6,:)=[real(16)::-v32(2), v32(1),      0,-x23(2), x23(1),      0]
 
-            b(1:3)= [real(8)::cross(x(:,1),v(:,1))-cross(x(:,2),v(:,2))]
-            b(4:6)= [real(8)::cross(x(:,2),v(:,2))-cross(x(:,3),v(:,3))]
+            b(1:3)= [real(16)::cross(x(:,1),v(:,1))-cross(x(:,2),v(:,2))]
+            b(4:6)= [real(16)::cross(x(:,2),v(:,2))-cross(x(:,3),v(:,3))]
             if(inv(a,6,work,ipiv))then
                b=matmul(a,b)
                write(*,*)sum(b(1:3))
@@ -97,10 +97,10 @@ contains
    logical function inv (b,n,work,ipiv) result(res)
       implicit none
       integer,intent(in)   :: n
-      real(8),intent(inout):: b(n,n)
-      real(8),intent(inout):: work(n)
+      real(16),intent(inout):: b(n,n)
+      real(16),intent(inout):: work(n)
       integer,intent(inout):: ipiv(n)
-      real(8) ::  c, d
+      real(16) ::  c, d
       integer :: i, j, k, m
       res=.true.
       do i=1,n
@@ -135,9 +135,9 @@ contains
    end function inv
 
    pure elemental subroutine swap(a,b)
-      real(8),intent(inout)::a
-      real(8),intent(inout)::b
-      real(8)::tmp
+      real(16),intent(inout)::a
+      real(16),intent(inout)::b
+      real(16)::tmp
       tmp=a;a=b;b=tmp
    end subroutine swap
 end program main
